LayoutTests/http/tests/websocket/tests/hybi/close-before-open.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test if Web Socket fires close event when WebSocket is opened and closed fore open event is received.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var urls = [
    "ws://127.0.0.1:8880/websocket/tests/hybi/echo",
    "ws://127.0.0.1:8880/websocket/tests/hybi/close-code-and-reason"
];

var testId = 0;

var test = function() {
    if (0 != testId)
        debug("closed on test " + testId);
    if (2 == testId)
        finishJSTest();
    else {
        ws = new WebSocket(urls[testId++]);
        ws.onclose = test;
        ws.close();
    }
};

test();

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
