LayoutTests/http/tests/websocket/tests/hybi/close-before-open-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/echo' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/close-code-and-reason' failed: WebSocket is closed before the connection is established.
Test if Web Socket fires close event when WebSocket is opened and closed fore open event is received.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

closed on test 1
closed on test 2
PASS successfullyParsed is true

TEST COMPLETE
